- Moved ASO class in a different package called ASOlibrary - Created AbstractLogicShape, ILogicShape, ShapeObject - ShapeObject: an object with a shape ... is very similar to concept of Layer.HasSize of PlayN ... When i will have a more clean code i could integrate it in AbstractSceneObject ... i must think about it! - Modified references to ASO* - Create AbstractSceneObjectBorder: An object border ... it's a proof ... - Added ShapeUtil that contains drawborders, BasicShapeInfo used by AbstractSceneObjectBorder

+++ forscene-core/src/forscene/core/entities/AbstractScene.java
@@ -13,7 +13,7 @@ public abstract class AbstractScene extends AbstractSceneObject{
 	private boolean USE_TIMER = false;
 	private boolean IS_READY_TO_SWITCH = false;
 	private boolean IS_CONDITIONAL = false;
-	private int seconds;
+	private float seconds;
 	
 	private AbstractScene next,prev;
 	
@@ -59,11 +59,11 @@ public abstract class AbstractScene extends AbstractSceneObject{
 	}
 
 	
-	public int getTimeout() {
+	public float getTimeout() {
 		return seconds;
 	}
 
-	public void setTimeout(int seconds) {
+	public void setTimeout(float seconds) {
 		this.seconds = seconds;
 		USE_TIMER = true;
 	}
